Implement customizable box inset

+++ src/main/java/mousemaster/Hint.java
@@ -2,7 +2,8 @@ package mousemaster;
 
 import java.util.List;
 
-public record Hint(int centerX, int centerY, int cellWidth, int cellHeight, List<Key> keySequence) {
+public record Hint(double centerX, double centerY, double cellWidth, double cellHeight,
+                   List<Key> keySequence) {
 
     boolean startsWith(List<Key> focusedHintKeySequence) {
         if (focusedHintKeySequence.isEmpty())
